+++ packages/stream_feeds/test/test_utils.dart
@@ -1,4 +1,7 @@
 export 'test_utils/event_types.dart';
 export 'test_utils/fakes.dart';
 export 'test_utils/mocks.dart';
+export 'test_utils/testers/activity_list_tester.dart';
+export 'test_utils/testers/activity_tester.dart';
+export 'test_utils/testers/feed_tester.dart';
 export 'test_utils/ws_test_helpers.dart';